ipc: Add basic requests
You can now communicate and gather data from the compositor using the library.
A simple test client located at `fht-compositor-ipc/test_client.rs` is there
for a code example of how to use the library.

For now you can ask for the compositor version and the outputs available
More to come!